Managed to draw upside down glyphs

+++ src/main/java/org/kurodev/kimage/kimage/font/glyph/Coordinate.java
@@ -1,8 +1,6 @@
 package org.kurodev.kimage.kimage.font.glyph;
 
-import org.kurodev.kimage.kimage.font.enums.GlyphFlag;
-
 import java.util.List;
 
-public record Coordinate(short x, short y, List<GlyphFlag> flags) {
+public record Coordinate(int x, int y, List<GlyphFlag> flags) {
 }
